Renamed summarizer_main to summarizer and classifier_main to classifier
